+++ src/App.tsx
@@ -2,8 +2,8 @@ import { Route, Routes, useRoutes } from "react-router-dom";
 import { useState, useEffect } from "react";
 import styled from "styled-components";
 import Home from "./HomePage/components/firstpage/Home";
-import Footer from "./components/Footer";
-import Navbar from "./HomePage/Nav";
+import Footer from "./HomePage/Footer";
+import Navbar from "./HomePage/components/navbar/Nav";
 import AboutUs from "./HomePage/components/aboutUs/AboutUs";
 import LoginForm from "./HomePage/LoginForm";
 import { IMember } from "./types/Member";
@@ -12,7 +12,7 @@ import FAQ from "./HomePage/FAQ";
 import auth from "./services/authService";
 import Logout from "./MyPages/Logout";
 import MyPage from "./MyPages/MyPage";
-import RenovetaForm from "./components/renovetaForm/RenovetaForm";
+import RenovetaForm from "./HomePage/components/renovetaForm/RenovetaForm";
 import RegisterForm from "./HomePage/RegisterForm";
 import Form from "./MyPages/Form";
 import { IUser } from "./types/User";
