+++ src/HomePage/LoginForm.tsx
@@ -3,6 +3,8 @@ import styled from "styled-components";
 import { useState } from "react";
 import useForm from "../common/Form";
 import auth from "../services/authService";
+import { ScrollToTop } from "../common/ScrollToTop";
+import { Link } from "react-router-dom";
 
 interface LoginFormData {
   email: string;
@@ -59,6 +61,10 @@ function LoginForm() {
       {renderInput({ label: "E-mail", name: "email" })}
       {renderInput({ label: "Password", name: "password", type: "password" })}
       {renderButton({ label: "Log in" })}
+
+      <Link to="/register" onClick={ScrollToTop} className="link">
+        Registrera dig här!
+      </Link>
     </Container>
   );
 }
@@ -66,6 +72,7 @@ function LoginForm() {
 export default LoginForm;
 
 const Container = styled.form`
+  display: grid;
   background-color: var(--bg-secondary);
   border: 5px solid var(--bg-primary);
   border-radius: 45px;
@@ -79,6 +86,14 @@ const Container = styled.form`
   margin-top: 20px;
   margin-bottom: 100px;
   margin-left: 400px;
+
+  .link {
+    margin-top: 10px;
+    font-size: 18px;
+    margin: 24px;
+    text-decoration: none;
+    cursor: pointer;
+  }
 
   @media screen and (max-width: 880px) {
     margin-left: 50px;
