+++ src/HomePage/components/calculator/BoxRight.tsx
@@ -1,36 +1,184 @@
 import { useState } from "react";
 import styled from "styled-components";
 import Button from "../../../common/Button";
+import Checkbox from "../../../common/Checkbox";
 import Input from "../../../common/Input";
-// import { Link } from "react-router-dom";
+import { Link } from "react-router-dom";
+import Questionmark from "../../../common/Questionmark";
+import { ScrollToTop } from "../../../common/ScrollToTop";
+
+enum RenovationType {
+  REROOFING = "Takomläggning",
+  PAINTING = "Målning",
+  CLEANING = "Tvätt",
+}
+
+enum MaterialType {
+  DEFAULT = "",
+  SHEET_METAL = "Plåt",
+  PAPER = "Papp/Shingel",
+  ETHERNITE = "Eternit",
+  BRICK_SHEETS = "Tegelpannor",
+  CONCRETE_SHEETS = "Betongpannor",
+}
 
 function BoxRight() {
-  const [input, setInput] = useState("kvm");
+  const [input, setInput] = useState<number>(0);
+  const [selectedType, setSelectedType] = useState<RenovationType | "">("");
+  const [selectedMaterial, setSelectedMaterial] = useState<MaterialType | "">(
+    ""
+  );
+
+  interface Data {
+    id: string;
+    name: MaterialType;
+    [RenovationType.REROOFING]: number;
+    [RenovationType.PAINTING]: number;
+    [RenovationType.CLEANING]: number;
+  }
+
+  const data: Data[] = [
+    {
+      id: "1",
+      name: MaterialType.SHEET_METAL,
+      [RenovationType.REROOFING]: 1150,
+      [RenovationType.PAINTING]: 350,
+      [RenovationType.CLEANING]: 180,
+    },
+    {
+      id: "2",
+      name: MaterialType.PAPER,
+      [RenovationType.REROOFING]: 750,
+      [RenovationType.PAINTING]: 350,
+      [RenovationType.CLEANING]: 180,
+    },
+    {
+      id: "3",
+      name: MaterialType.ETHERNITE,
+      [RenovationType.REROOFING]: 1850,
+      [RenovationType.PAINTING]: 350,
+      [RenovationType.CLEANING]: 180,
+    },
+    {
+      id: "4",
+      name: MaterialType.BRICK_SHEETS,
+      [RenovationType.REROOFING]: 1500,
+      [RenovationType.PAINTING]: 350,
+      [RenovationType.CLEANING]: 180,
+    },
+    {
+      id: "5",
+      name: MaterialType.CONCRETE_SHEETS,
+      [RenovationType.REROOFING]: 1200,
+      [RenovationType.PAINTING]: 350,
+      [RenovationType.CLEANING]: 180,
+    },
+  ];
+
+  function getTotalPrice() {
+    let total = 0;
+
+    const material = data.find((item) => item.name === selectedMaterial);
+
+    if (!material) return total;
+
+    total += material[selectedType as RenovationType] * input;
+
+    return total;
+  }
+
+  interface TextQuestions {
+    questionOne: string;
+    questionTwo: string;
+    questionThree: string;
+  }
+
+  const text: TextQuestions = {
+    questionOne:
+      "Takrenovering är ...Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+    questionTwo: "Material till ditt tak ...",
+    questionThree: "Räkna ut ditt tak area..",
+  };
+
   return (
     <Right>
       <div>
-        <p>1. Vad är det för typ av renovering?</p>
+        <QuestionContainer>
+          <a>1. Vad är det för typ av takrenovering?</a>
+          <Questionmark text={text.questionOne} />
+        </QuestionContainer>
+
+        <span>
+          {Object.values(RenovationType).map((option, i) => {
+            return (
+              <StyledCheckBox key={i}>
+                <Checkbox
+                  className="checkbox-input"
+                  key={i}
+                  id={i.toString()}
+                  checked={option === selectedType}
+                  handleCheck={(e) =>
+                    setSelectedType(e.target.name as RenovationType)
+                  }
+                  name={option}
+                  type="radio"
+                />
+              </StyledCheckBox>
+            );
+          })}
+        </span>
       </div>
       <div>
-        <p>
-          2. Hur stort är ditt tak i kvm?
-          <Input
-            className="input"
-            value={input}
-            name="kvm"
-            type="text"
-            onChange={(event: any) => setInput(event.target.value)}
-          />
-        </p>
+        <QuestionContainer>
+          <a>2. Vilket material består ditt tak av?</a>
+          <Questionmark text={text.questionTwo} />
+        </QuestionContainer>
+
+        <span>
+          {data.map((option) => {
+            return (
+              <StyledCheckBox key={option.id}>
+                <Checkbox
+                  className="checkbox-input"
+                  key={option.id}
+                  id={option.id.toString()}
+                  checked={option.name === selectedMaterial}
+                  handleCheck={(e) =>
+                    setSelectedMaterial(e.target.name as MaterialType)
+                  }
+                  name={option.name}
+                  type="radio"
+                />
+              </StyledCheckBox>
+            );
+          })}
+        </span>
       </div>
       <div>
-        <p>3. Vilket material består ditt tak av?</p>
+        <QuestionContainer>
+          <a>3. Hur stort är ditt tak i kvm?</a>
+          <Questionmark text={text.questionThree} />
+        </QuestionContainer>
+
+        <Input
+          className="input"
+          value={input}
+          name="kvm"
+          type="text"
+          onChange={(e) => setInput(Number(e.target.value))}
+        />
       </div>
-      <p>4. Preferenser</p>
       <End>
-        <Button primary={false} type="submit" label={"Beräkna"} />
-        <p>Jag vill veta mer om renoveta</p>
-        {/* <Link to={"/aboutus"}>Jag vill veta mer om renoveta</Link> */}
+        <a>{getTotalPrice()}kr</a>
+        <p>Prisindikation redovisas med arbete och material</p>
+        <Link onClick={ScrollToTop} to="/renovetaform">
+          <Button
+            className="button"
+            primary={false}
+            type="submit"
+            label={"Kom igång!"}
+          />
+        </Link>
       </End>
     </Right>
   );
@@ -39,7 +187,7 @@ function BoxRight() {
 export default BoxRight;
 
 const Right = styled.span`
-  width: 55%;
+  width: 650px;
   display: grid;
   background-color: var(--bg-primary);
   color: var(--text-secondary);
@@ -48,8 +196,9 @@ const Right = styled.span`
   border: 5px solid var(--bg-secondary);
   padding-left: 72px;
   line-height: 28px;
+  padding-top: 50px;
 
-  p {
+  a {
     font-size: 25px;
     font-weight: 700;
     font-style: bold;
@@ -59,15 +208,48 @@ const Right = styled.span`
   input {
     border: 1.5px solid var(--bg-secondary) !important;
   }
+
+  span {
+    display: grid;
+    grid-template-columns: auto auto;
+    margin: 10px;
+  }
+
+  @media screen and (max-width: 1000px) {
+    width: 450px;
+    font-size: 12px;
+    padding-top: 20px;
+    padding-left: 32px;
+
+    a {
+      font-size: 16px;
+    }
+
+    .button {
+      width: 120px;
+      height: 30px;
+    }
+  }
+`;
+
+const QuestionContainer = styled.div`
+  display: flex;
+  @media screen and (max-width: 1000px) {
+    font-size: 14px;
+  }
+`;
+
+const StyledCheckBox = styled.div`
+  width: 150px;
+  margin-bottom: 5px;
+  border-radius: 5px;
+
+  .checkbox-input {
+    margin-right: 5px;
+  }
 `;
 
 const End = styled.div`
-  padding-left: 20%;
-
-  p {
-    font-style: bold;
-    font-weight: 700;
-    font-size: 14;
-    padding-left: 21;
-  }
+  text-align: center;
+  margin-bottom: 20px;
 `;
