--- src/HomePage/components/calculator/Header.tsx
@@ -1,13 +1,0 @@
-function Header() {
-  return (
-    <div>
-      <h1>Testa Renoveta kalkylen</h1>
-      <p>
-        Undrar du vad din renovering kommer att kosta, <br /> behövs kanske
-        bygglov och hur är det egentligen med rot-avdrag?
-      </p>
-    </div>
-  );
-}
-
-export default Header;
