+++ src/HomePage/components/firstpage/Home.tsx
@@ -6,7 +6,7 @@ import { IList } from "./howRenovetaworks/IList";
 import Review from "./Review";
 import WhatisRenoveta from "./WhatisRenoveta";
 import Compilation from "./Compilation";
-import Calculator from "../../../components/calculator/Calculator";
+import Calculator from "../calculator/Calculator";
 
 function Home() {
   const [list, setList] = useState<IList[]>([]);
